Clear search instead of filling in result.
This behaviour makes sense both when you want to search for different players one after the other, and when clicking on a different player name in the games list.
It is worse than the previous behaviour only when you want to modify your last search.
Comment or revert if you don't like this change.